+++ src/AngleSharp.Diffing.Tests/Strategies/AttributeStrategies/AttributeComparerTest.cs
@@ -27,7 +27,7 @@ namespace AngleSharp.Diffing.Strategies.AttributeStrategies
             var comparison = ToAttributeComparison(@"<b foo>", "foo",
                                                     "<b foo>", "foo");
 
-            AttributeComparer.Compare(comparison, CompareResult.Different).ShouldBe(CompareResult.Same);
+            AttributeComparer.Compare(comparison, CompareResult.Unknown).ShouldBe(CompareResult.Same);
         }
 
         [Fact(DisplayName = "When two attributes does not have the same name, the compare result is Different")]
@@ -36,7 +36,7 @@ namespace AngleSharp.Diffing.Strategies.AttributeStrategies
             var comparison = ToAttributeComparison(@"<b foo>", "foo",
                                                     "<b bar>", "bar");
 
-            AttributeComparer.Compare(comparison, CompareResult.Different).ShouldBe(CompareResult.Different);
+            AttributeComparer.Compare(comparison, CompareResult.Unknown).ShouldBe(CompareResult.Different);
         }
 
         [Fact(DisplayName = "When two attribute values are the same, the compare result is Same")]
@@ -45,7 +45,7 @@ namespace AngleSharp.Diffing.Strategies.AttributeStrategies
             var comparison = ToAttributeComparison(@"<b foo=""bar"">", "foo",
                                                    @"<b foo=""bar"">", "foo");
 
-            AttributeComparer.Compare(comparison, CompareResult.Different).ShouldBe(CompareResult.Same);
+            AttributeComparer.Compare(comparison, CompareResult.Unknown).ShouldBe(CompareResult.Same);
         }
 
         [Fact(DisplayName = "When two attribute values are different, the compare result is Different")]
@@ -54,7 +54,7 @@ namespace AngleSharp.Diffing.Strategies.AttributeStrategies
             var comparison = ToAttributeComparison(@"<b foo=""bar"">", "foo",
                                                    @"<b foo=""baz"">", "foo");
 
-            AttributeComparer.Compare(comparison, CompareResult.Different).ShouldBe(CompareResult.Different);
+            AttributeComparer.Compare(comparison, CompareResult.Unknown).ShouldBe(CompareResult.Different);
         }
 
         [Fact(DisplayName = "When the control attribute is postfixed with :ignoreCase, " +
@@ -64,7 +64,7 @@ namespace AngleSharp.Diffing.Strategies.AttributeStrategies
             var comparison = ToAttributeComparison(@"<b foo:ignoreCase=""BAR"">", "foo:ignorecase",
                                                    @"<b foo=""bar"">", "foo");
 
-            AttributeComparer.Compare(comparison, CompareResult.Different).ShouldBe(CompareResult.Same);
+            AttributeComparer.Compare(comparison, CompareResult.Unknown).ShouldBe(CompareResult.Same);
         }
 
         [Fact(DisplayName = "When the control attribute is postfixed with :regex, " +
@@ -75,7 +75,7 @@ namespace AngleSharp.Diffing.Strategies.AttributeStrategies
             var comparison = ToAttributeComparison(@"<b foo:regex=""foobar-\d{4}"">", "foo:regex",
                                                    @"<b foo=""foobar-2000"">", "foo");
 
-            AttributeComparer.Compare(comparison, CompareResult.Different).ShouldBe(CompareResult.Same);
+            AttributeComparer.Compare(comparison, CompareResult.Unknown).ShouldBe(CompareResult.Same);
         }
 
         [Theory(DisplayName = "When the control attribute is postfixed with :regex:ignoreCase " +
@@ -90,7 +90,7 @@ namespace AngleSharp.Diffing.Strategies.AttributeStrategies
             var comparison = ToAttributeComparison($@"<b {controlAttrName}=""foobar-\d{{4}}"">", controlAttrName,
                                                    @"<b foo=""FOOBAR-2000"">", "foo");
 
-            AttributeComparer.Compare(comparison, CompareResult.Different).ShouldBe(CompareResult.Same);
+            AttributeComparer.Compare(comparison, CompareResult.Unknown).ShouldBe(CompareResult.Same);
         }
     }
 }